% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPdata.R
\docType{data}
\name{CPdata}
\alias{CPdata}
\title{simulated data}
\format{
a data set containig 500 observations of 6 variables.
}
\source{
simulated data
}
\usage{
data(CPdata)
}
\description{
This data file contains 500 simulated subjects of 1 response variable (Y) and
5 predictors (X1-X5). They are generated with underlying subgroup structures.
}
