% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPlane.R
\name{CPlane}
\alias{CPlane}
\title{Change Plane method for subgroup identification.}
\usage{
CPlane(
  data.tr,
  data.te = NULL,
  yind,
  xind,
  zind,
  ini.theta = rep(1, length(zind)),
  tol = 0.001,
  K = 10
)
}
\arguments{
\item{data.tr}{a matrix or data frame of the training dataset.}

\item{data.te}{a matrix or data frame of the testing dataset.}

\item{yind}{the column number or column name of the response variable.}

\item{xind}{a numeric or character vector, containing the column numbers or column names of the predictors (covariates).}

\item{zind}{a numeric or character vector, containing the column numbers or column names of the threshold variables.}

\item{ini.theta}{a numeric vector, containing the initial values of change plane parameters (the linear combination coefficients
of the threshold variables). For better performance of this function, it is highly recommended that the user
provide an input based on preknowledge or a reasonable guess of the change plane structure.
Default is ini.theta=rep(1, length(zind)).}

\item{tol}{error tolerance in parameter estimation and optimization. Default is tol=1e-3.}

\item{K}{maximum time of iterations in parameter estimation and optimization. Dault is K=10.}
}
\value{
A list consisting of the following components:
\item{train.res}{estimation results for the training dataset, including estimated response value (Yhat), subgroup id (subgroup), mean squared error (mse), estimated change plane parameter (theta), thresholds on the linear combination of threshold variables (threshold) and the segment regression coefficients (coefficient).}
\item{test.res}{estimation results for the testing dataset, including estimated response value (Yhat), subgroup id (subgroup) and mean squared error (mse).}
}
\description{
Idenfiy subgroups marked by multiple parallel change planes.
}
\details{
This function identifies subgroups and makes predictions by fitting a multithreshold change plane regression
model to the given training dataset. Subgroups are thus charaterised by a linear combination of the specified
threshold variables. The change plane parameters, thresholds and regression coefficients are estimated by a
two stage approach proposed by Li et al (2018). Note that this function should only be used when there are
multiple threshold variables so that a linear combination of them makes sense, i.e. length(zind) > 1. Otherwise,
one should turn to the CPoint() function instead.
}
\examples{
out <- CPlane(data.tr=CPdata[1:350,], data.te=CPdata[-(1:350),], yind = 1, xind = 2:6, zind = 2:6,
ini.theta = c(sqrt(0.5), -sqrt(0.5), 0, 0, 0))
}
\references{
Li, J., Y. Li, B. Jin, and M. R. Kosorok (2021). Multithreshold change plane model: estimation theory and applications in subgroup identification. Statistics in Medicine 40(15), 3440-3459.
}
