% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPoint.R
\name{CPoint}
\alias{CPoint}
\title{Change Point method for subgroup identification.}
\usage{
CPoint(data.tr, data.te = NULL, yind, xind, zind, c = seq(0.5, 1.5, 0.1))
}
\arguments{
\item{data.tr}{a matrix or data frame of the training dataset.}

\item{data.te}{a matrix or data frame of the testing dataset.}

\item{yind}{the column number or column name of the response variable.}

\item{xind}{a numeric or character vector, containing the column numbers or column names of the predictors (covariates).}

\item{zind}{the column number or column name of the threshold variable.}

\item{c}{a numeric vector to determine the initial segment length in the splitting stage of TSMCD, i.e. the tentative choices of initial segment length m = c * sqrt(n), where n is the sample size of the training dataset. Default is c = seq(0.5, 1.5, 0.1)}
}
\value{
A list consisting of the following components:
\item{train.res}{estimation results for the training dataset, including estimated response value (Yhat), subgroup id (subgroup), mean squared error (mse), detected change points (threshold) and the segment regression coefficients (coefficient).}
\item{test.res}{estimation results for the testing dataset, including estimated response value (Yhat), subgroup id (subgroup) and mean squared error (mse).}
}
\description{
Subgroup identification by change point detection on a (single) specified threshold variable Z.
}
\details{
This function identifies subgroups and makes predictions by fitting a change point (threshold regression) 
model to the given training dataset. The change points (thresholds) and regression coefficients 
are estimated by a two stage multiple change-points detection (TSMCD) method proposed by Li and Jin (2018).
}
\examples{
out <- CPoint(data.tr=CPdata[1:350,], data.te=CPdata[-(1:350),], yind=1, xind=2:6, zind=2)
}
\references{
Li, J. and B. Jin (2018). Multi-threshold accelerated failure time model. The Annals of Statistics 46(6A), 2657-2682.
}
