% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPointMA.R
\name{CPointMA}
\alias{CPointMA}
\title{Change Point Model Average method for subgroup identification.}
\usage{
CPointMA(
  data.tr,
  data.te = NULL,
  yind,
  xind,
  zind,
  c = seq(0.5, 1.5, 0.1),
  penalty = c("SCAD", "MCP", "LASSO")
)
}
\arguments{
\item{data.tr}{a matrix or data frame of the training dataset.}

\item{data.te}{a matrix or data frame of the testing dataset.}

\item{yind}{the column number or column name of the response variable.}

\item{xind}{a numeric or character vector, containing the column numbers or column names of the predictors (covariates).}

\item{zind}{a numeric or character vector, containing the column numbers or column names of the threshold variables.}

\item{c}{a numeric vector to determine the initial segment length in the splitting stage of TSMCD for submodel estimation, i.e. the tentative choices of initial segment length m = c * sqrt(n), where n is the sample size of the training dataset. Default is c = seq(0.5, 1.5, 0.1)}

\item{penalty}{the penalty to be used in the model average step, including 'SCAD', 'MCP' and 'LASSO'. Default is 'SCAD'.}
}
\value{
A list consisting of the following components:
\item{train.res}{estimation results for the training dataset, including estimated response value (Yhat), subgroup id (subgroup), mean squared error (mse), detected change points (threshold) and the segment regression coefficients (coefficient).}
\item{test.res}{estimation results for the testing dataset, including estimated response value (Yhat), subgroup id (subgroup) and mean squared error (mse).}
\item{submodel.res}{estimation results for each of the submodels.}
}
\description{
While CPoint() identifies subgroups based on a single threshold variable, this function 
detects change points each threshold variable separately and then assembles the subgrouping 
results through model averaging.
}
\details{
The methodology of this function consists of two levels. An individual change point (threshold regression) 
submodel is fitted based on each of the given threshold variables in the training dataset. 
Then their model averaging weights are estimated so that a weighted ensemble of these submodels 
can be used to further approximate the true model. Note that this function should only be used
when there are multiple threshold variables, i.e. length(zind) > 1. Otherwise, one should turn to
the CPoint() function instead.
}
\examples{
out <- CPointMA(data.tr=CPdata[1:350,], data.te=CPdata[-(1:350),], yind=1, xind=2:6, zind=2:6)
}
\references{
Li, J. and B. Jin (2018). Multi-threshold accelerated failure time model. The Annals of Statistics 46(6A), 2657-2682.
}
